django/requirements.txt
arrow==1.2.2
asgiref==3.5.0
blessed==1.19.1
boto3==1.21.23
botocore==1.24.23
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coverage==6.2
cryptography==36.0.2
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-extensions==3.1.3
django-fernet-encrypted-fields==0.1.1
django-filter==2.4.0
django-jazzmin==2.5.0
django-picklefield==3.0.1
django-q==1.3.9
django-sequences==2.6
django-storages==1.12.3
djangorestframework==3.13.1
ecdsa==0.17.0
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
minio==7.1.2
Pillow==9.0.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
python-dateutil==2.8.2
python-jose==3.3.0
python-keycloak==0.26.1
pytz==2022.1
redis==3.5.3
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==5.3.0
